Split off `UTxOStatistics` to separate module.
